+++ IaC/bin/ia_c.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import * as cdk from 'aws-cdk-lib';
-import { AuthStack } from '../lib/ia_c-stack';
+import { AuthStack } from '../lib/stack/cloud_tunes-stack';
 
 const app = new cdk.App();
