+++ src/artists/models/artist-model.d.ts
