+++ src/artists/models/artist-model.js
