+++ src/artists/models/artist-model.ts
