+++ src/artists/services/artist-service.ts
@@ -1,0 +1,6 @@
+import type { Artist } from "../models/artist-model.ts";
+
+export async function createArtist(artist: Artist): Promise<void> {
+    console.log(artist)
+    return
+}
