+++ src/contents/models/content-models.js
