+++ src/home/components/Home.tsx
@@ -2,7 +2,7 @@ import {useEffect} from "react";
 import {useNavigate} from "react-router-dom";
 import {globalSignOut} from "../../users/services/login-service.ts";
 import {TokenStorage} from "../../users/services/user-token-storage-service.ts";
-import {UploadContent} from "../../content/components/UploadContent.tsx";
+import {UploadContent} from "../../contents/components/UploadContent.tsx";
 
 export const Home = () => {
     useEffect(() => {
