+++ src/users/models/aws-calls.ts
@@ -15,31 +15,3 @@ export interface CognitoTokens {
     idToken: string;
     refreshToken: string;
 }
-
-
-export interface UploadContentRequest {
-    filename: string;
-    filetype: string;
-    filesize: number;
-    title: string;
-    imageUrl?: string;
-    albumId?: string;
-    albumName?: string;
-    genres: string[];
-    artistIds: string[];
-}
-
-export interface UploadContentResponse {
-    contentId: string;
-    filename: string;
-    filetype: string;
-    filesize: number;
-    title: string;
-    imageUrl?: string;
-    albumId?: string;
-    albumName?: string;
-    genres: string[];
-    artistIds: string[];
-    createdAt: string;
-    updatedAt: string;
-}
