--- src/users/services/artist-service.ts
@@ -1,5 +1,0 @@
-import type { Artist } from "../models/artist-model";
-
-export async function createArtist(artist: Artist): Promise<void> {
-    return 
-}
